Switch to using m4 bolts everywhere
The main issue was the size of button head m4 bolts resulting in much larger adjacent holes that were nearly the diameter of the side_tab_width. Fixed this by parameterizing the side tabs to only occupy a fraction of the available space so that there can be more material left on either side when the bolt head holes are cut out. Fixes #6.

There's still not too much material left between the bottom of the bolt head holes and the bottom of the enclosure (0.9 mm), so it may be worth extending the height of the enclosure in the future, though that would be backwards incompatible with previous enclosures.