Add slot for bolt to act as flap backstop
Allow for a bolt on the right side to stick inward and act as a flap backstop. I'm mostly guessing at the positioning and also made it a slot rather than a hole for the bolt to hopefully allow for some adjustment. Fixes #5 (but will reopen if it doesn't work out in practice).