Initial python implementation of controller
Implemented new integer-based prototype of the splitflap module controller to avoid slow floating point math on the Arduino. Prototyped in python for easier testing and debugging.

The main difference is that instead of tracking the position as a floating point flap-index (i.e. [0.0-40.0) for the 40 flaps) we now track everything in terms of integer motor steps, taking care to preserve the ability to handle tricky gear ratios like the ~63.684:1 commonly found in the 5V version of the 28byj-48 stepper. The other change is to do any "complex" math upfront (when commanding the module to move to a particular flap) so that the incremental updates can simply do a few integer comparisons to drive the state machine.

Also added a global configuration parameter in this design to easily enable/disable home calibration; while home calibration is essential to robust operation, it's helpful to be able to do initial tests of a flap module without worrying about the home sensor.

Started porting the new implementation to c++.