Add option for rendering just the front flaps and unique letters per module
Change render_flaps to support multiple levels, with a new intermediate option that only renders the front top/bottom flap. This mades the model significantly smaller and render faster, which is useful if you want to preview a large number of modules at once and don't need all 40 flaps rendered on every single module.

Also allowed for different characters to be displayed on each module when rendering multiple modules.