Rough initial PCB layout
Starting to lay out components on the new design. Mostly addresses #10 - the board outline has been shifted to start at 20,20 (with 0.1mm margin on each edge) so coordinates should be a bit more sane now. Moved mounting holes to be a part of the GP2S60 footprint to guarantee their relative spacing stays compatible with the enclosure dimensions.