Inline SplitflapModule methods
Move the entirety of SplitflapModule to the .h file so that all methods can be inlined, for better performance. When driving all 12 modules at once, the overhead of method calls slows down the control loop frequency, so timing becomes less accurate and the stepper motors are more prone to skipping due to inconsistent step widths.

Also applied a minor change to the startup LED animation, and conditions under which we refresh the status LEDs.