Enable panelized 2d output for bulk laser-cutting
Add panelization support to the projection renderer so that it can lay out the elements in a location for proper panelization when rendering for SVG output. Improved the previous N^2 redundant line removal to first bucket by rough slope-intersect values so that the N^2 comparisons only need to happen within a small bucket rather than over all lines in the entire document (which quickly becomes too slow).

The design is panelized 2 per row (rotated 180 degrees and back-to-back).

The `generate_2d.py` script now takes an optional `--panelize <number>` argument to generate a panelized SVG with the specified number of copies of the design. Since redundant-line-removal happens after rendering and panelization, it will remove lines shared between duplicate copies when panelized.