Accurately model 3mm flap notch depth, without changing other dimensions
As discussed in #24, the model was erroneously rendering the flap notch as `thickness` (3.2mm) when in reality the recommended flap notch punch is only 3mm wide.

This introduces a new `flap_notch_depth`, set to 3mm, and updates the spool dimensions, etc to be based on the flap width minus the flap_notch_depth on each side. However, in order to keep the model's dimensions consistent with the previous version (for which numerous protypes have been made that work well), the flap_width_slop was reduced from 0.5mm to 0.1mm. This keeps the spool exactly the same size, but allows the model to accurately reflect changes to the `flap_notch_depth`. I also added an assert to warn you if you accidentally make the `flap_notch_depth` too large such that the flaps would be wider than the spool (and therefore interfere with the flap window and enclosure).