Add "basic IO" mode for using ULN2003A motor drivers without custom PCB
Updates #32. Now the Arduino code can be configured for directly controlling the "standard" ULN2003A-based 28BYJ-48 motor drivers instead of using the more complex shift-register based drivers.

To do so, set `NUM_MODULES` to 3 or fewer, `NEOPIXEL_DEBUGGING_ENABLED` to `false`, and `SPI_IO` to `false`. The motors and sensors can be connected to the pins described in basic_io_config.h