Simplify and strengthen inter-module connector
Fixes #36. This eliminates the locking "tab" on the inter-module connector and instead makes the connector bracket slide into a slot in the front edge of the left/right side panels. Once the front panel is installed, the connector is locked in place so there's no need for the complexity of a locking tab.