Wait until schematic plot exists with timeout
Instead of always waiting for 2 seconds. Typically plotting the
schematic does not take that long so the 2 seconds wait makes plotting
the schematic take way longer than it should.

Is now also give a clearer error message when plotting the
schematic fails since the script should not attempt to read a
non-existing file when plotting times out.