Fix oval drills in panelize.py
Fixes #60. Panelize.py was expecting just a single `size` argument for pad drills, but for slots it can be something like `(drill oval 1 3)`, so the oval dimensions were lost when panelized. This just replaces the explicit `size` field and allows `drill` specs to have any number of args that are copied as-is.