Round motor window and CAD tweaks (#88)
Rounds the corners of the motor window (r=5) for slightly more stiffness and a nicer fit + finish. With a radius of 5 mm this should still be a loose fit around the motor (at least according to the model). On the flat panel the spool retaining wall was moved to the X center of the window to clear the rounded corners while still maintaining its shared bottom edge.

This also adjusts the lasercut panel so that the module connector pieces won't cut material out of the enclosure's left side. To do this the top and bottom pieces were moved down and the connectors were nested between the tabs in the top right corner. This keeps within the existing boundaries of the panel so it uses no extra material.

For what it's worth these new connector positions break down and start intersecting the other parts around a thickness of ~5, though the connector geometry itself seem to break down there as well.

Last but not least this removes the `center=true` arguments on `circle()` calls, which aren't needed and were creating warnings in OpenSCAD (2019.05).

---

* Round motor mounting window

Adds a 5 mm radius on each corner, and moves the spool retaining wall to the X center on the flat panel to clear the corner

* Fix OpenSCAD warnings

'center' is not a parameter for circle (OpenSCAD 2019.05)

* Remove unused flat layout variable (cad)

* Comment additional flat layout sections

* Move top/bottom pieces to bottom of panel

Frees up some extra room on the top right for the connector brackets

* Move flat connector brackets to top corner

Prevents having to cut them out of the left side

* Position flat connectors off of each-other

Shares slightly more of an edge for less cutting distance and packs pieces closer together.

* Motor window round: improve func. readability

No functional change.