Parse SVG viewboxes instead of paths
Smarter and much more efficient implementation of the viewbox recalculation (e75f57e). Compares the existing viewbox and imported viewbox and returns a new viewbox enclosing both of them.

This calculation is done in the 'import' function rather than calling 'fix_dimensions' separately, as the viewbox only changes when new paths are imported.