Font Rework and SVG Generation (#92)
This PR overhauls how fonts are handled in the SCAD assembly, and adds a font generation feature for rendering SVG font files for cutting. The goal is to make it easy to generate custom split-flap fonts for desktop cutting machines (Cricut, Silhouette, etc.).

Changes:
* Added parametric font generation to SVG, for cutting your own letters (`generate_fonts.py` and `font_generator.scad`)
* Letter rendering is now handled by a single function, with an argument to trim and position for either top or bottom 
* Letter geometry / position now matches flap geometry, so both `flap()` and `flap_letter()` can be called from the same point without additional transformations
* Letters are now trimmed to flap outlines exactly, rather than using a 2D exclusion area for the opposite flap
* Fixed vertical text alignment and baseline position. Characters like ',' and '.' now sit at the baseline, and characters like 'Q' and 'j' can extend below it
* Letters are now parameterized (`flap_fonts.scad`):
    * Font selection
    * Letter scaling (X/Y)
    * Letter position offsets (X/Y from center)
* Letter position can now optionally compensate for the gap between flaps (`letter_gap_comp`)

Also included:
* In 3D mode, the spool assembly can now be hidden (`render_spool`)

With these new options I've tweaked the default letter scaling width (0.8 of previous) so that letters such as 'M' don't exceed the flap bounds. I've also nudged the X and Y offsets so the letters are better centered with the Roboto font.

### Font Generation

`font_generator.scad` renders a string of text across an X/Y array of flaps. This uses the same flap and font settings as the main assembly and can be customized to change the font, scaling, position offsets, and more. 

The Python script allows overriding most of the parameters from the command line, so font generation can be automated if need be. The flap outlines are purposefully exported for alignment purposes.