CAD Render Colors (#105)
A number of changes to the assembly to make it easy to adjust the colors.

First, the assembly color variations are now generated from a single base color. This means you can change the one 'assembly' color and still keep the color variation between parts. The scalars were calculated from the current values so everything should keep the same appearance as before. The etch color, similarly, has been broken out into a variable and can now be changed.

Second, I've added variables for the hardware (nuts + bolts) and flap color. This allows you to change out the steel hardware for black oxide or swap the flaps from white to black. The letter color is calculated as the inverse of the flap color, so white flaps get black letters and black flaps get white letters (and red flaps get cyan letters and so on and so forth).

There should be no functional changes to the design.