Flap Case (#103)
While I was waiting for the enclosure to be lasercut I got a head start on applying the vinyl to the flaps, but had nowhere to put them. So I designed a 3D printed case that holds a full set of 40 flaps.

The design includes a finger hole at the bottom for pushing flaps out, a 'pinch' cutout on the sides for grabbing flaps, and a rounded slot on the bottom for retaining a rubber band or other strap to keep the flaps in place. I've printed eight of these (from the Fusion 360 version of the design) and they work great.

The OpenSCAD version submitted is fully modular and will scale in height for different flap counts. I also added an array feature so you can create a single part with multiple rows and columns of containers, such as a 1x4 version for all of the modules on a single driver.