Hall Effect Sensor Spacing and Jig (#101)
This PR dynamically adjusts the hall effect sensor's height off of the PCB in the model based on its distance from the spool. The sensor will maintain the same relative distance from the spool as the model is changed and throw an error if its distance from the PCB exceeds the max length of the sensor's leads (per the datasheet). I also created a small lasercut jig that fits between the PCB and the sensor to aid in soldering.

Using your v0.5 sensor assembly video as reference, I set the spool offset to 0.7 mm which corresponds to a distance of 5.8 mm from the PCB. Note that this PR has a conflict with #100 that will need to be resolved, as the new jig is cut out of the motor window area.