Sensor PCB Case (#104)
As with the flap case (#103), while I'm waiting for the enclosures to be finished I have a bunch of assembled sensor PCBs sitting around. This is a little dangerous since the hall effect sensors stick out from the board and can be accidentally bent out of position. So I designed a 3D printed case for them to safely sit in until the modules can be assembled.

The design has a `cases_per_row` parameter for generating multiple cases in one design, as well as a `dual_rows` parameter that mirrors a second row along the back for double the number of PCB spots: