Spool Fitment Adjustments (#107)
This PR makes a few minor adjustments to how the spool assembly fits together.

First and most importantly, the `spool_tab_clearance` value, which affects the fitment between the spool struts and the spool end pieces, has been changed from a slight clearance / free fit (`0.1 mm`) to a slight interference fit (`-0.02 mm`). This was an oversight - the spool, as designed, needs to be an interference fit otherwise the spool endpieces fall right off. Mea culpa.

The spool struts themselves, as well as the spool retaining wall, will keep their `0.1 mm` clearance value so they'll fit together smoothly. This also gives you a little bit of wiggle room when trying to slot the spool struts into the interference-fit endcaps, which will undoubtedly be a little tricky with an acrylic enclosure.

This also very slightly reduces the interference fit with the motor shaft (from `-1.0` to `-0.8`). With an accurately-cut acrylic enclosure the motor shaft is a very tight fit and reliably cracks the spool's mounting point when sliding it on. I'm cautious about reducing this too much, but it's clear that the value as-is is too snug on the spindle.