CAD: Remove equilateral centering for triangles
This is clever in principle but poor in practice, because you have to assume the intention behind the centering command. Do you want to center based on the size (bounding box), or do you want to center based on the rotation if the triangle is equilateral? Easier to always assume bounding box centering and get rid of the trig.