Laser Layout Mirror Option (#106)
This adds the option to mirror the 2D laser-cut design so that the outside faces are facing 'up' / towards the laser when cut. This is useful for materials with a special finish on only one side like Ponoko's matte black acrylic, which is cut with the matte finish facing towards the laser.

The mirroring can be set with the render_2d_mirror flag, or with the --mirror option in the generate_2d.py script.

With the addition of the alignment indicators (PR #111) I think we've reached the point where it no longer makes sense to mirror the individual etched portions. Instead, I've updated the etch function to prevent rendering the laser etches if the mirror option is set.

I've also updated the `generate_2d.py` and `projection_renderer.py` scripts to pass OpenSCAD components that don't return geometry instead of raising an exception, and added an additional `--no-etch` option for exporting the unmirrored design without the etched portions.

#### Demos:

* [Standard](https://user-images.githubusercontent.com/24282108/102972173-a65a0f00-44c8-11eb-83fd-8175ecb2810f.png)
* [No Etching (`--no-etch` option)](https://user-images.githubusercontent.com/24282108/102972169-a3f7b500-44c8-11eb-9ecc-2451aa41b5bf.png)
* [Mirrored (`--mirror` option)](https://user-images.githubusercontent.com/24282108/102972164-a22df180-44c8-11eb-9067-4f7faf6c30c5.png)