CAD: Move flap thickness slop to spool slop
This increase was a mistake. The difference in thickness (0.2 on either side) should have been added to the *spool* slop instead of the flap width slop, as the inner width of the spool (flap_width_slop) was not affected by the thickness parameter, while the outer width of the spool (spool_width_slop) *was*. This commit corrects that oversight.

This reverts commit a68fc988bc73c146c28ed1dc2c8e02c990d5ab93 (squashed as part of 7be69dbeb4fb7c1aa4b29027eb3e49a9b121c6a2).