CAD: Move flap thickness slop to spool slop (#114)
This increase was a mistake. The difference in thickness (0.2 on either side) should have been added to the *spool* slop instead of the flap width slop, as the inner width of the spool (flap_width_slop) was not affected by the thickness parameter, while the outer width of the spool (spool_width_slop) *was*. This commit corrects that oversight.

This reverts commit a68fc988bc73c146c28ed1dc2c8e02c990d5ab93 (squashed as part of 7be69dbeb4fb7c1aa4b29027eb3e49a9b121c6a2).

On my test display the spool will sometimes stall while rapiding. It appears that this is caused by imperfect assembly of the spool and the flaps sliding sideways, causing them to get caught on the upper corner of the front panel. I've traced the issue back to this change and I'm hopeful this PR will fix the issue.

Nudges this from 4 mil to 6 mil, just for that little extra clearance to avoid binding. (And for peace of mind)