CAD: Revert flap width slop narrowing
The flaps *must* move freely without binding, otherwise the display does not work. Given the variability of online laser-cutting services, tolerance stacking, and the imprecision of hand assembling the 7 separate components that make up the spool, it seems imprudent to reduce this value without being able to test the results firsthand.

Some narrowing here would be beneficial, but it's difficult to say exactly how far to go before binding will occur on outlier assemblies. Put another way: a display that may sometimes stall due to misalignment is preferrable to a display that does not 'flip' at all.

This partially reverts commit f4e7714.