CAD: Increase Motor Clearance (#121)
Increases the clearance between the motor chassis and the right outside wall of the previous module. Also refactors this variable from 'slop' to 'clearance' for a better description of what it represents.

This is still playing a bit of catch-up from the thickness change (#93). In the previous release where the thickness parameter was set to 3.2 mm but the manufactured panels were actually 3.0 mm thick, there was extra space between the motor chassis and the previous module. This change adds that additional 0.4 mm of clearance back as part of the motor clearance variable.

The additional clearance also makes the widths between the two designs approximately equal - `enclosure_width` is 82.2 mm on v0.6, and 82.25 mm here. (The discrepancy is the additional 0.05 of flap width slop that was added in #114.) Note that #122 bumps this further to 82.35.