Module Connector Case (#126)
Since the connector pieces are a little fragile on there own and are not necessarily used for each module I designed a 3D printable case to store them. The design is modular and can be extended based on:

* The number of connectors per stack
* The number of stacks per row
* The number of rows per design (1 or 2)

Each stack of connectors can be retained with a rubber band. I've printed my own 3x2 version and it works great.