Enclosure Alignment Bar (#142)
Creates a hole in the left enclosure panel that is aligned with the back edge of the right side of the enclosure, to be used with a dowel rod or other round stock as an alignment aid for multiple modules.

The connector pieces are a bit of a pain to use since you have to remove the front panels and deal with the captive nuts that can be hidden underneath a full spool of flaps. This creates an alternate way to align the back edge of the modules without extending the length of the right panel and increasing the size and cost of the laser-cut design. In larger displays it also allows modules to be swapped out without having to undo any of the hardware.

Note that this does *not* prevent the module from twisting forwards about the alignment hole (along Z), and it would need either an accompanying strip on the front or a bracket attached to the right panel as well. For example, a pivoting latch attached to the mounting bolt or a thin tie around the bottom enclosure piece.