Center front face, tabs, and bolts relative to the window (#159)
Previously the front face and tabs were all centered between the left and right walls, but this resulted in features looking oddly off-center when viewed from the front (where the walls aren't visible and the only other visible feature is the window).

This changes the tabs and front face positioning so that everything is centered relative to the front window. The tabs are made slightly smaller, and the top/bottom pieces are no longer horizontally symmetric as a result. This is not a breaking change (overall dimensions are all the same), but this could not be placed to the _right_ of older modules since the front face has shifted and the motor sticks out further relative to the edge of that face.