Elecrow v2 export fixes (#239)
PDFs generated for elecrow on the latest master currently look like
this:

![image](https://github.com/scottbez1/splitflap/assets/15990762/8a188939-1e57-4fbf-9400-bb90268b4982)



Dimensions appear to be placed incorrectly when the svg output is
mirrored. I updated `add_dimensions` to account for mirroring when its
enabled.

I tested this in a 20.04 docker container setup similarly to the actions
environment and the output looks correct. Result is below:


[elecrow_dim_fix.pdf](https://github.com/user-attachments/files/15907775/elecrow_dim_fix.pdf)

When testing locally, I also discovered that the inkscape cli has
deprecated `--verbs` in favor of `--actions` in version 1.2+. This PR
adds a check to handle both inkscape < 1.2 and >= 1.2 when generating
elecrow pdfs